// libs/features/medical-board/src/lib/shared/enums/document-transaction-id.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum DocumentTransactionId {
  MTN_MB_DOCTOR = 'MTN_MED_BOARD_DOCTOR',
  MTN_MB_CONTRACT_MODIFY = 'MTN_MED_BOARD_MEMBER_CONTRACT_MODIFY',
  MTN_MB_DOCTOR_TERMINATE = 'MTN_MED_BOARD_MEMBER_CONTRACT_TERMINATE',
  MTN_MB_TERMINATE = 'MTN_MED_BOARD_MEMBER_CONTRACT_TERMINATE'
  //MTN_MED_BOARD_MEMBER_CONTRACT_TERMINATE = 'MTN_MED_BOARD_MEMBER_CONTRACT_TERMINATE'
}