// libs/features/medical-board/src/lib/shared/enums/medical-board-wizard-types.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

/* Form wizard types enum */

export enum MedicalBoardWizardTypes {
  PERSONAL_DETAILS = 'MEDICAL-BOARD.PERSONAL-DETAILS',
  CONTRACT_DETAILS = 'MEDICAL-BOARD.CONTRACT-DETAILS',
  DOCUMENTS = 'MEDICAL-BOARD.DOCUMENTS'
}