// libs/features/medical-board/src/lib/shared/enums/nationality-category-enum.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum NationalityCategoryEnum {
  SAUDI_PERSON = 'Saudi_Person',
  NON_SAUDI = 'Non_Saudi_Person',
  GCC_PERSON = 'GCC_Person',
  NIN_ID = 'NIN',
  IQAMA_ID = 'Iqama',
  GCC_ID = 'GCCId',
  BORDER_NO = 'Border No',
  IQAMA_NO = 'iqamaNo'
}